fix: swap lambda entries
